Split testbench.py into smaller modules.
This is part of the changes for #1187. It moves the classes and
functions to separate modules. The routing functions are still all in
the main testbench.py module, and it is rather large, we will fix that
in a future PR.
